# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Transcription API
  version: 1.0.0
  description: |
    Contract for the local transcription service.
    Error payloads follow the shared ErrorResponse schema.
servers:
  - url: http://localhost:8000
    description: Local development

tags:
  - name: Health
  - name: Upload
  - name: Transcriptions
  - name: Audio

paths:
  /api/health:
    get:
      tags: [Health]
      operationId: getHealth
      summary: Health check
      responses:
        '200':
          description: Service is healthy.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                ok:
                  externalValue: ./examples/happy/health-200.json
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  externalValue: ./examples/error/health-500.json

  /api/upload:
    post:
      tags: [Upload]
      operationId: uploadAudio
      summary: Upload an audio file and return transcription
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Audio file to transcribe.
                language:
                  type: string
                  default: en
                  description: Optional language hint passed to transcription service.
                source:
                  type: string
                  enum: [recording, upload]
                  default: upload
                  description: Declares whether audio came from live recording or file upload.
      responses:
        '201':
          description: Upload processed and transcription completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
              examples:
                success:
                  externalValue: ./examples/happy/upload-201.json
        '400':
          description: Invalid file type or malformed request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidFileType:
                  externalValue: ./examples/error/upload-400-invalid-file-type.json
        '413':
          description: Uploaded file exceeds configured size limit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                fileTooLarge:
                  externalValue: ./examples/error/upload-413-file-too-large.json
        '422':
          description: Audio processing failed before transcription.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '502':
          description: Failed to process/transcribe audio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                processingFailed:
                  externalValue: ./examples/error/upload-500-processing-failed.json

  /api/transcriptions:
    get:
      tags: [Transcriptions]
      operationId: listTranscriptions
      summary: List transcriptions
      parameters:
        - in: query
          name: limit
          required: false
          description: Number of items to return.
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: offset
          required: false
          description: Number of items to skip.
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated list of transcriptions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscriptionListResponse'
              examples:
                success:
                  externalValue: ./examples/happy/transcriptions-200.json
        '400':
          description: Reserved for explicit domain-level pagination failures.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidPagination:
                  externalValue: ./examples/error/transcriptions-400-invalid-pagination.json
        '422':
          description: Query parameter validation failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Unexpected server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  externalValue: ./examples/error/transcriptions-500.json

  /api/transcriptions/{id}:
    get:
      tags: [Transcriptions]
      operationId: getTranscription
      summary: Get transcription by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Transcription identifier.
      responses:
        '200':
          description: Transcription found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transcription'
              examples:
                success:
                  externalValue: ./examples/happy/transcription-by-id-200.json
        '404':
          description: Transcription not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  externalValue: ./examples/error/transcription-by-id-404.json

  /api/audio/{id}:
    get:
      tags: [Audio]
      operationId: getAudio
      summary: Download stored audio file by transcription id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Transcription identifier.
      responses:
        '200':
          description: Raw audio stream/download.
          headers:
            Content-Type:
              description: Media type of the stored audio file.
              schema:
                type: string
                example: audio/wav
            Content-Disposition:
              description: Suggests download filename.
              schema:
                type: string
                example: attachment; filename="meeting.wav"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
              examples:
                sampleBytes:
                  externalValue: ./examples/happy/audio-200.bin.base64.txt
        '404':
          description: Audio file/transcription id not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  externalValue: ./examples/error/audio-404.json

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, version]
      properties:
        status:
          type: string
          enum: [ok]
        version:
          type: string

    Chunk:
      type: object
      required: [start, end, text]
      properties:
        start:
          type: number
          minimum: 0
        end:
          type: number
          minimum: 0
        text:
          type: string

    TranscriptionSummary:
      type: object
      required:
        - id
        - filename
        - source
        - language
        - status
        - text
        - created_at
        - audio_url
      properties:
        id:
          type: string
        filename:
          type: string
        source:
          type: string
          enum: [recording, upload]
        language:
          type: string
        duration_seconds:
          type: number
          nullable: true
        status:
          type: string
          enum: [completed, failed]
        text:
          type: string
        created_at:
          type: string
          format: date-time
        audio_url:
          type: string

    Transcription:
      allOf:
        - $ref: '#/components/schemas/TranscriptionSummary'
        - type: object
          properties:
            chunks:
              type: array
              items:
                $ref: '#/components/schemas/Chunk'
          required: [chunks]

    TranscriptionListResponse:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TranscriptionSummary'
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0

    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true

    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string